feat(priority-grid): refactor the big userPrioritizationGrid into seperate files
